requirements.txt
absl-py==2.3.0
accelerate==1.7.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
args==0.1.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
blinker==1.9.0
bpy==4.0.0
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
clint==0.5.1
clip==1.0
colorama==0.4.6
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
datasets==3.6.0
dds-cloudapi-sdk==0.5.1
decorator==5.2.1
decord==0.6.0
deepface==0.0.93
defusedxml==0.7.1
diffusers==0.33.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dreamsim==0.2.1
einops==0.8.1
evo==1.31.1
exceptiongroup==1.3.0
executing==2.2.0
facexlib==0.3.0
fairscale==0.4.13
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.14.0
filterpy==1.4.5
fire==0.4.0
firerequests==0.1.4
flash-attn==2.7.2.post1
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.40.3
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.2
google-genai==1.5.0
google-pasta==0.2.0
googleapis-common-protos==1.71.0
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
groundingdino==0.1.0
grpcio==1.73.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
hpsv2==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.32.4
hydra-core==1.3.2
icecream==2.1.4
identify==2.6.12
idna==3.10
image-reward==1.5
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
iopath==0.1.10
ipdb==0.13.13
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
keras==2.15.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
liger_kernel==0.5.4
llvmlite==0.44.0
lmdb==1.6.2
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
mmcv==2.2.0
mmengine==0.8.4
model-index==0.1.11
modelscope==1.28.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.4.4
multiprocess==0.70.16
namex==0.1.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.2
numexpr==2.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
omegaconf==2.3.0
open_clip_torch==2.32.0
openai==1.86.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openxlab==0.1.2
opt_einsum==3.4.0
optree==0.17.0
ordered-set==4.1.0
orjson==3.10.18
oss2==2.17.0
packaging==24.2
pandas==2.3.0
pandas-flavor==0.8.1
parso==0.8.4
patsy==1.0.2
peft==0.15.2
pexpect==4.9.0
pillow==11.1.0
pingouin==0.5.5
pip==25.1
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.1.1
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
pyee==11.1.1
Pygments==2.19.1
pyiqa==0.1.13
pyparsing==3.2.3
pyppeteer==2.0.0
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2023.4
PyYAML==6.0.2
quickxorhash==1.0.5
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.28.2
requests-oauthlib==2.0.0
resample2d-cuda==0.0.0
retina-face==0.0.17
rich==13.4.2
roma==1.5.4
rosbags==0.10.11
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.11.13
safehttpx==0.1.6
safetensors==0.4.5
SAM-2==1.0
scikit-image==0.25.2
scikit-learn==1.7.0
scikit-video==1.1.11
scipy==1.15.3
seaborn==0.13.2
segment-anything==1.0
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
sentry-sdk==2.32.0
setuptools==60.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spaces==0.37.1
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.6
supervision==0.25.1
sympy==1.14.0
t2v_metrics==1.2
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==1.0.13
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trimesh==4.7.1
triton==3.0.0
typer==0.16.0
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.34.3
validators==0.35.0
virtualenv==20.31.2
wcwidth==0.2.13
webdataset==0.2.111
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.1
xarray==2025.6.1
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zipp==3.22.0
zstandard==0.23.0
